requirements.txt
# Citation
# Date: 04/30/2024
# Generated running 'pip freeze > requirements.txt' as instructed in the Flask Starter App
# Source URL: https://github.com/osu-cs340-ecampus/flask-starter-app/?tab=readme-ov-file#requirementstxt
netCDF4
matplotlib
pvlib
pandas
datetime
pytz
siphon
ruff
aiohttp
asyncio
pprintjson
openmeteo-requests
requests-cache
retry-requests
numpy
python-dotenv
click
Flask
Flask-WTF
Flask-Security-Too
Flask-Login
Flask-Migrate
gunicorn
itsdangerous
Jinja2
MarkupSafe
mysqlclient
flask-mysqldb
Werkzeug
black
python-decouple